convert file to ts and change to useMutation
